Add 'save poll' and refactor
Refactor add poll / cancel poll so that they are contained in the
NewPoll component and make callbacks to IndexPage for actions.

TODO: consider whether the actual actions should take place in NewPoll
and only callback to IndexPage be to let it know which elements should
be visible.

Add 'save poll' functionality

TODO: make the list of polls render properly; believe this should be
in another component ...