Complete initial rollout
Add data visualisation using chart.js and react-chartjs

TODO: getting a warning that React.create (used in react-chartjs) is
being deprecated in the next version of React. Need to consider whether
this is an issue (may never update this app) but hopefully the package
will be updated by it's maintainers.